go06: part2: check if new paths leads to loops
We make a naive check that if there's an obstacle that we previously
turned at, in the turn direction, then we can start new paths.

This solves the example input.

This is however still very naive. We might need to check if we need to
bounce on multiple obstacle before we come onto a path we've already
been at.

And indeed 491 is too low for the real input data.